ok finally solved the mask issue.
But i'll still leave the new mask code, it should still be more robust

+++ utils_/utils.py
@@ -84,7 +84,7 @@ def imputations_acc_justy(model, dloader, device):
     with torch.no_grad():
         for i, data in enumerate(dloader, 0):
             image, ids, DOY, x_categ, x_cont, y_gts = data[0].to(device), data[1].to(device).type(torch.float32), data[2].to(
-                device).type(torch.float32), data[3].to(device).type(torch.float32), data[4].type(torch.LongTensor).to(
+                device).type(torch.float32), data[3].to(device).type(torch.float32), data[4].type(torch.float32).to(
                 device),data[5].to(device,dtype=torch.long)#,data[6].to(device).type(torch.float32)
             _, x_categ_enc, x_cont_enc, con_mask = embed_data_mask(x_categ, x_cont, model, vision_dset, DOY=DOY)
             reps = model.tab_net.transformer(x_categ_enc, x_cont_enc, con_mask)
@@ -155,7 +155,7 @@ def class_wise_acc_(model, dloader, device):
     with torch.no_grad():
         for i, data in enumerate(dloader, 0):
             image, ids, DOY, x_categ, x_cont, y_gts = data[0].to(device), data[1].to(device).type(torch.float32), data[2].to(
-                device).type(torch.float32), data[3].to(device).type(torch.float32), data[4].type(torch.LongTensor).to(
+                device).type(torch.float32), data[3].to(device).type(torch.float32), data[4].type(torch.float32).to(
                 device),data[5].to(device, dtype=torch.long)#,data[6].to(device).type(torch.float32)
             _, x_categ_enc, x_cont_enc, con_mask = embed_data_mask(x_categ, x_cont, model, vision_dset, DOY=DOY)
             reps = model.tab_net.transformer(x_categ_enc, x_cont_enc, con_mask)
@@ -212,7 +212,7 @@ def mean_sq_error(model, dloader, device, vision_dset):
     with torch.no_grad():
         for i, data in enumerate(dloader, 0):
             image, ids, DOY, x_categ, x_cont, y_gts = data[0].to(device), data[1].to(device).type(torch.float32), data[2].to(
-                device).type(torch.float32), data[3].to(device).type(torch.float32), data[4].type(torch.LongTensor).to(
+                device).type(torch.float32), data[3].to(device).type(torch.float32), data[4].type(torch.float32).to(
                 device),data[5].to(device, dtype=torch.long)#,data[6].to(device).type(torch.float32)
             _, x_categ_enc, x_cont_enc, con_mask = embed_data_mask(x_categ, x_cont, model, vision_dset, DOY=DOY)
             reps = model.tab_net.transformer(x_categ_enc, x_cont_enc, con_mask)
